requirements.txt
agents==1.4.0
google_api_python_client==2.177.0
isodate==0.7.2
langchain_core==0.3.72
langchain_openai==0.3.28
langchain_text_splitters==0.3.9
pandas==2.3.1
pydantic==2.11.7
python-dotenv==1.1.1
pytubefix==9.4.1
streamlit==1.47.1
streamlit_player==0.1.5
youtube_transcript_api==1.2.2
